std: Move fs/path to the top-level (#3100)
