+++ std/fs/mod.ts
@@ -5,8 +5,7 @@ export * from "./ensure_file.ts";
 export * from "./ensure_link.ts";
 export * from "./ensure_symlink.ts";
 export * from "./exists.ts";
-export * from "./glob.ts";
-export * from "./globrex.ts";
+export * from "./expand_glob.ts";
 export * from "./move.ts";
 export * from "./copy.ts";
 export * from "./read_file_str.ts";
