--- std/fs/path.ts
@@ -1,3 +1,0 @@
-// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-export * from "./path/mod.ts";
-export * from "./path/interface.ts";
