+++ std/installer/mod.ts
@@ -2,9 +2,9 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 const { env, stdin, args, exit, writeFile, chmod, run } = Deno;
 import { parse } from "../flags/mod.ts";
-import * as path from "../fs/path.ts";
 import { exists } from "../fs/exists.ts";
 import { ensureDir } from "../fs/ensure_dir.ts";
+import * as path from "../path/mod.ts";
 
 const encoder = new TextEncoder();
 const decoder = new TextDecoder("utf-8");
