+++ std/io/util.ts
@@ -2,8 +2,9 @@
 const { Buffer, mkdir, open } = Deno;
 type File = Deno.File;
 type Reader = Deno.Reader;
+import * as path from "../path/mod.ts";
 import { encode } from "../strings/mod.ts";
-import * as path from "../fs/path.ts";
+
 // `off` is the offset into `dst` where it will at which to begin writing values
 // from `src`.
 // Returns the number of bytes copied.
