+++ std/path/README.md
@@ -1,0 +1,7 @@
+# Deno Path Manipulation Libraries
+
+Usage:
+
+```ts
+import * as path from "https://deno.land/std/path/mod.ts";
+```
