+++ std/path/globrex.ts
