+++ std/path/interface.ts
