+++ std/path/mod.ts
@@ -23,3 +23,8 @@ export const format = path.format;
 export const parse = path.parse;
 export const sep = path.sep;
 export const delimiter = path.delimiter;
+
+export { EOL, SEP, SEP_PATTERN, isWindows } from "./constants.ts";
+export * from "./interface.ts";
+export * from "./glob.ts";
+export * from "./globrex.ts";
