+++ std/prettier/main_test.ts
@@ -1,10 +1,9 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-import { join } from "../fs/path.ts";
-import { EOL } from "../fs/path/constants.ts";
 import { assertEquals } from "../testing/asserts.ts";
 import { test, runIfMain } from "../testing/mod.ts";
-import { xrun } from "./util.ts";
 import { copy, emptyDir } from "../fs/mod.ts";
+import { EOL, join } from "../path/mod.ts";
+import { xrun } from "./util.ts";
 const { readAll, execPath } = Deno;
 
 const decoder = new TextDecoder();
